mipmap, exec: two rail decisions that a both-rails build would have made wrong
Both of these were `cfg` pairs that a build carrying both features does not
merely leave ambiguous — it resolves them *silently and incorrectly*, which is
what turning both on was going to reveal:

* `mipmap.rs` declared `let chain = ...` twice, once per arm. With both
  features on, the Vulkan binding shadows the Metal one, so `generate_via_metal`
  runs, its result is dropped on the floor, and every guest `generateMipmaps`
  quietly takes the CPU box filter. The compiler's only complaint is an unused
  variable.
* `exec.rs` declared `let translation_pending` twice the same way, so the Vulkan
  preflight would run and its answer be replaced by `false` — a packet with cold
  AIR still translating would be executed rather than deferred, which is the
  duplicate-clear/fence/dispatch replay the deferral exists to prevent.

`Backend::generate_mipmap_chain` returns `MipmapGeneration`, and the three cases
are three because they want three different handlings: `Chain` is the filtered
result, `Unfiltered` sends the caller to the shared box filter (same levels,
slower — never a loss), and `Refused` must reach the guest as a refusal, because
a rail that *has* a filtered path and was rejected by it must not be quietly
approximated. The `NoDevice`-declines-but-everything-else-refuses rule moves
into the Metal rail with the error type it reads, along with the
`mipmap_metal_fallback` line, since only the rail knows the reason.

`Backend::preflight_translations` is a safety answer, not an optimisation:
`false` promises the packet can be executed to completion now. The default is
that promise for a rail that compiles no shader ahead of the record using one.
The Vulkan halves move to `runtime/exec/vulkan.rs`. Its `ExecPhase::Preflight`
timing stays on the runtime side and now runs unconditionally, so a rail that
preflights nothing shows as the zero it costs rather than as an absent column
the leftover `Header` absorbs.

With these two, `cargo check` of the **library** with both features on has no
errors — the remaining co-build work is `select()` and three test files.

Verified on both arms: feature matrix, clippy `-D warnings`, full suites
serially (metal 1327, vulkan 2269, support 351, unchanged).
